requirements.txt
aenum==3.1.0
androguard==3.3.5
appnope==0.1.2
asn1crypto==1.4.0
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
cycler==0.10.0
decorator==5.0.9
Flask==1.1.1
future==0.18.2
googleplayapi==1.1
gpsoauth==1.0.0
greenlet==1.1.1
gremlinpython==3.4.4
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
influxdb-client==1.21.0
ipython==7.27.0
isodate==0.6.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
kiwisolver==1.3.2
lxml==4.4.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
msgpack==1.0.2
networkx==2.6.2
numpy==1.17.4
parsel==1.6.0
parso==0.8.2
pexpect==4.8.0
pg8000==1.21.1
pickleshare==0.7.5
Pillow==8.3.2
priority==1.3.0
prompt-toolkit==3.0.20
Protego==0.1.16
protobuf==3.12.2
ptyprocess==0.7.0
publicsuffix==1.1.1
publicsuffixlist==0.6.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
PyDispatcher==2.0.5
pydot==1.4.2
Pygments==2.10.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest-httpserver==0.3.4
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
queuelib==1.6.2
requests==2.26.0
scramp==1.4.1
Scrapy==2.5.0
Scrapy-UserAgents==0.0.1
sentry-sdk==0.13.5
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.0
tornado==4.5.3
tqdm==4.62.2
traitlets==5.1.0
treq==21.5.0
Twisted==19.10.0
urllib3==1.26.6
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==2.0.1
zope.interface==5.4.0
